+++ src/components/form/textarea/TextArea.stories.tsx
@@ -6,12 +6,19 @@ const meta: Meta = {
     title: "Components/Form/TextArea",
     component: TextArea,
     tags: ["autodocs", "form"],
+    argTypes: {
+        error: { type: "boolean" },
+        disabled: { type: "boolean" },
+        readOnly: { type: "boolean" },
+    },
 };
 
 type Story = StoryObj<typeof TextArea>;
 
 const Primary: Story = {
-    args: {},
+    args: {
+        placeholder: "Type something here",
+    },
 };
 
 export {