+++ src/components/form/textarea/TextArea.styled.ts
@@ -12,7 +12,6 @@ const StyledTextArea = styled("textarea", {
     height: "unset",
     resize: "vertical",
     minHeight: dimensionsPxToRem(103),
-
 });
 
 export {